+++ CHANGELOG.md
@@ -7,11 +7,21 @@ See also the [GitHub releases page](https://github.com/FriendsOfSymfony/FOSHttpC
 ------------------
 
 * Support Symfony 4.
+
+### Testing
+
+* Upgraded phpunit to 5.7 / 6. If you use anything from the
+  `FOS\HttpCache\Test` namespace you need to update your project to use
+  PHPUnit 6 (or 5.7, if you are using PHP 5.6).
+
+### Symfony HttpCache
+
+* Added a `PurgeTagsListener` for tag based invalidation with the Symfony
+  `HttpCache` reverse caching proxy. This requires the newly created
+  [Toflar Psr6Store](https://github.com/Toflar/psr6-symfony-http-cache-store)
+  built on PSR-6 cache and supporting pruning expired cache entries.
 * Using Request::isMethodCacheable rather than Request::isMethodSafe to
   correctly handle OPTIONS and TRACE requests.
-* Upgraded phpunit to 5.7 / 6. If you use anything from the FOS\HttpCache\Test
-  namespace you need to update your project to use phpunit 6 (or 5.7, if you
-  are using PHP 5.6).
 
 2.0.2
 -----